talk-endee/requirements.txt
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.0.1
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.4.2
msgpack==1.0.7
